requirements.txt
django==5.0.3
djangorestframework==3.15.1
django-environ==0.11.2
django-filter==24.2
factory-boy==3.3.0
psycopg2==2.9.9